refactor: remove mutable rules subclass initialization
